Also do the conversion in `build_fronted_http_client()`.
