+++ nym-vpn-core/crates/nym-vpn-api-client/src/client.rs
@@ -14,6 +14,7 @@ use time::OffsetDateTime;
 
 use crate::{
     error::{Result, VpnApiClientError},
+    fronted_http_client::api_url_to_url,
     request::{
         ApplyFreepassRequestBody, CreateAccountRequestBody, CreateSubscriptionKind,
         CreateSubscriptionRequestBody, RegisterDeviceRequestBody, RequestZkNymRequestBody,
@@ -60,9 +61,10 @@ impl VpnApiClient {
         user_agent: UserAgent,
         resolver_overrides: Option<&ResolverOverrides>,
     ) -> Result<Self> {
-        let url = Self::api_url_to_url(&api_url)?;
+        let url = api_url_to_url(&api_url)?;
+        let has_front = url.has_front();
 
-        let mut builder = nym_http_api_client::Client::builder(url.clone())
+        let mut builder = nym_http_api_client::Client::builder(url)
             .map_err(Box::new)
             .map_err(VpnApiClientError::CreateVpnApiClient)?;
 
@@ -70,7 +72,7 @@ impl VpnApiClient {
             .with_user_agent(user_agent.clone())
             .with_timeout(NYM_VPN_API_TIMEOUT);
 
-        if url.has_front() {
+        if has_front {
             builder = builder.with_fronting(FrontPolicy::OnRetry);
         }
 
@@ -137,7 +139,7 @@ impl VpnApiClient {
 
         let urls: Vec<Url> = api_urls
             .iter()
-            .map(Self::api_url_to_url)
+            .map(api_url_to_url)
             .collect::<Result<Vec<_>, _>>()?;
 
         let has_front = urls.iter().any(|url| url.has_front());
@@ -1350,34 +1352,6 @@ impl VpnApiClient {
             .await
             .map_err(Box::new)
             .map_err(VpnApiClientError::GetVpnNetworkDetails)
-    }
-
-    fn api_url_to_url(api_url: &ApiUrl) -> Result<Url, VpnApiClientError> {
-        let parse_url = |s: &str| -> Result<url::Url, VpnApiClientError> {
-            match url::Url::parse(s) {
-                Ok(url) => Ok(url),
-                Err(_) => {
-                    let with_scheme = format!("http://{s}");
-                    url::Url::parse(&with_scheme)
-                        .map_err(|_e| VpnApiClientError::InvalidUrl { url: s.to_string() })
-                }
-            }
-        };
-
-        let url: url::Url = parse_url(&api_url.url)?;
-        let fronts: Option<Vec<url::Url>> = api_url
-            .front_hosts
-            .as_ref()
-            .map(|hosts| {
-                hosts
-                    .iter()
-                    .map(|host| parse_url(host))
-                    .collect::<Result<Vec<_>, _>>()
-            })
-            .transpose()?;
-        Url::new(url, fronts).map_err(|_e| VpnApiClientError::InvalidUrl {
-            url: api_url.url.to_string(),
-        })
     }
 }
